Potential fix for code scanning alert no. 3: Wrong type of arguments to formatting function

+++ server/src/debug.c
@@ -103,7 +103,7 @@ void locid_log(char *str, ...)
 }
 
 int locid_ssl_err_cb(const char *str, size_t len, void *u) {
-	locid_log("%s: %.*s",(char *)u,len-1,str);
+	locid_log("%s: %.*s", (char *)u, (int)(len-1), str);
 	return(0);
 }
